Final Update

+++ backend-br/controllers/testing.js
@@ -1,7 +1,8 @@
 const Reminder = require("../models/Reminder");
 
 const deleteAllReminders = async (req, res) => {
-  const deleteRes = await Reminder.deleteMany({});
+  const { userId } = req.params;
+  const deleteRes = await Reminder.deleteMany({ createdBy: userId });
   console.log(deleteRes);
 
   res.status(200).json({ msg: "Successfully deleted all Reminders." });
